requirements.txt
matplotlib==3.7.0
Pillow==9.4.0
python-telegram-bot==20.3
sympy==1.11.1
